[build] [js] [scss] Add uniform build steps for SASS and JS.
