+++ assets/css/admin/checkout.css
@@ -1,1 +1,1 @@
-@media screen and (max-width: 782px){#wpbody-content{padding-bottom:0 !important}}/*# sourceMappingURL=checkout.css.map */
+@media screen and (max-width:782px){#wpbody-content{padding-bottom:0!important}}
