+++ assets/js/nojquery.ba-postmessage.js
@@ -25,116 +25,4 @@
  * 5. The use of `$.isFunction` is replaced by a functionally-identical check.
  * 6. The `$:nomunge` YUI option is no longer necessary.
  */
-
-function NoJQueryPostMessageMixin(postBinding, receiveBinding) {
-
-    var setMessageCallback, unsetMessageCallback, currentMsgCallback,
-        intervalId, lastHash, cacheBust = 1;
-
-  if (window.postMessage) {
-
-    if (window.addEventListener) {
-      setMessageCallback = function(callback) {
-        window.addEventListener('message', callback, false);
-      }
-
-      unsetMessageCallback = function(callback) {
-        window.removeEventListener('message', callback, false);
-      }
-    } else {
-      setMessageCallback = function(callback) {
-        window.attachEvent('onmessage', callback);
-      }
-
-      unsetMessageCallback = function(callback) {
-        window.detachEvent('onmessage', callback);
-      }
-    }
-
-    this[postBinding] = function(message, targetUrl, target) {
-      if (!targetUrl) {
-        return;
-      }
-
-      // The browser supports window.postMessage, so call it with a targetOrigin
-      // set appropriately, based on the targetUrl parameter.
-      target.postMessage( message, targetUrl.replace( /([^:]+:\/\/[^\/]+).*/, '$1' ) );
-    }
-
-    // Since the browser supports window.postMessage, the callback will be
-    // bound to the actual event associated with window.postMessage.
-    this[receiveBinding] = function(callback, sourceOrigin, delay) {
-      // Unbind an existing callback if it exists.
-      if (currentMsgCallback) {
-        unsetMessageCallback(currentMsgCallback);
-        currentMsgCallback = null;
-      }
-
-      if (!callback) {
-        return false;
-      }
-
-      // Bind the callback. A reference to the callback is stored for ease of
-      // unbinding.
-      currentMsgCallback = setMessageCallback(function(e) {
-        switch(Object.prototype.toString.call(sourceOrigin)) {
-        case '[object String]':
-          if (sourceOrigin !== e.origin) {
-            return false;
-          }
-          break;
-        case '[object Function]':
-          if (sourceOrigin(e.origin)) {
-            return false;
-          }
-          break;
-        }
-
-        callback(e);
-      });
-    };
-
-  } else {
-
-    this[postBinding] = function(message, targetUrl, target) {
-      if (!targetUrl) {
-        return;
-      }
-
-      // The browser does not support window.postMessage, so set the location
-      // of the target to targetUrl#message. A bit ugly, but it works! A cache
-      // bust parameter is added to ensure that repeat messages trigger the
-      // callback.
-      target.location = targetUrl.replace( /#.*$/, '' ) + '#' + (+new Date) + (cacheBust++) + '&' + message;
-    }
-
-    // Since the browser sucks, a polling loop will be started, and the
-    // callback will be called whenever the location.hash changes.
-    this[receiveBinding] = function(callback, sourceOrigin, delay) {
-      if (intervalId) {
-        clearInterval(intervalId);
-        intervalId = null;
-      }
-
-      if (callback) {
-        delay = typeof sourceOrigin === 'number'
-          ? sourceOrigin
-          : typeof delay === 'number'
-            ? delay
-            : 100;
-
-        intervalId = setInterval(function(){
-          var hash = document.location.hash,
-            re = /^#?\d+&/;
-          if ( hash !== lastHash && re.test( hash ) ) {
-            lastHash = hash;
-            callback({ data: hash.replace( re, '' ) });
-          }
-        }, delay );
-      }
-    };
-
-  }
-
-  return this;
-}
+function NoJQueryPostMessageMixin(n,e){var t,i,o,s,a,r=1;return window.postMessage?(window.addEventListener?(t=function(n){window.addEventListener("message",n,!1)},i=function(n){window.removeEventListener("message",n,!1)}):(t=function(n){window.attachEvent("onmessage",n)},i=function(n){window.detachEvent("onmessage",n)}),this[n]=function(n,e,t){e&&t.postMessage(n,e.replace(/([^:]+:\/\/[^\/]+).*/,"$1"))},this[e]=function(n,e,s){if(o&&(i(o),o=null),!n)return!1;o=t((function(t){switch(Object.prototype.toString.call(e)){case"[object String]":if(e!==t.origin)return!1;break;case"[object Function]":if(e(t.origin))return!1}n(t)}))}):(this[n]=function(n,e,t){e&&(t.location=e.replace(/#.*$/,"")+"#"+ +new Date+r+++"&"+n)},this[e]=function(n,e,t){s&&(clearInterval(s),s=null),n&&(t="number"==typeof e?e:"number"==typeof t?t:100,s=setInterval((function(){var e=document.location.hash,t=/^#?\d+&/;e!==a&&t.test(e)&&(a=e,n({data:e.replace(t,"")}))}),t))}),this}
