+++ assets/js/postmessage.js
@@ -1,135 +1,1 @@
-(function ($, undef) {
-    var global = this;
-
-    // Namespace.
-    global.FS = global.FS || {};
-
-    global.FS.PostMessage = function ()
-    {
-        var
-            _is_child = false,
-            _postman = new NoJQueryPostMessageMixin('postMessage', 'receiveMessage'),
-            _callbacks = {},
-            _base_url,
-            _parent_url = decodeURIComponent(document.location.hash.replace(/^#/, '')),
-            _parent_subdomain = _parent_url.substring(0, _parent_url.indexOf('/', ('https://' === _parent_url.substring(0, ('https://').length)) ? 8 : 7)),
-            _init = function () {
-                _postman.receiveMessage(function (e) {
-                    var data = JSON.parse(e.data);
-
-                    if (_callbacks[data.type]) {
-                        for (var i = 0; i < _callbacks[data.type].length; i++) {
-                            // Execute type callbacks.
-                            _callbacks[data.type][i](data.data);
-                        }
-                    }
-                }, _base_url);
-            },
-            _hasParent = ('' !== _parent_url),
-            $window = $(window),
-            $html = $('html');
-
-        return {
-            init : function (url, iframes)
-            {
-                _base_url = url;
-                _init();
-
-                // Automatically receive forward messages.
-                FS.PostMessage.receiveOnce('forward', function (data){
-                    window.location = data.url;
-                });
-
-                iframes = iframes || [];
-
-                if (iframes.length > 0) {
-                    $window.on('scroll', function () {
-                        for (var i = 0; i < iframes.length; i++) {
-                            FS.PostMessage.postScroll(iframes[i]);
-                        }
-                    });
-                }
-            },
-            init_child : function ()
-            {
-                this.init(_parent_subdomain);
-
-                _is_child = true;
-
-                // Post height of a child right after window is loaded.
-                $(window).bind('load', function () {
-                    FS.PostMessage.postHeight();
-
-                    // Post message that window was loaded.
-                    FS.PostMessage.post('loaded');
-                });
-            },
-            hasParent : function ()
-            {
-                return _hasParent;
-            },
-            postHeight : function (diff, wrapper) {
-                diff = diff || 0;
-                wrapper = wrapper || '#wrap_section';
-                this.post('height', {
-                    height: diff + $(wrapper).outerHeight(true)
-                });
-            },
-            postScroll : function (iframe) {
-                this.post('scroll', {
-                    top: $window.scrollTop(),
-                    height: ($window.height() - parseFloat($html.css('paddingTop')) - parseFloat($html.css('marginTop')))
-                }, iframe);
-            },
-            post : function (type, data, iframe)
-            {
-                console.debug('PostMessage.post', type);
-
-                if (iframe)
-                {
-                    // Post to iframe.
-                    _postman.postMessage(JSON.stringify({
-                        type: type,
-                        data: data
-                    }), iframe.src, iframe.contentWindow);
-                }
-                else {
-                    // Post to parent.
-                    _postman.postMessage(JSON.stringify({
-                        type: type,
-                        data: data
-                    }), _parent_url, window.parent);
-                }
-            },
-            receive: function (type, callback)
-            {
-                console.debug('PostMessage.receive', type);
-
-                if (undef === _callbacks[type])
-                    _callbacks[type] = [];
-
-                _callbacks[type].push(callback);
-            },
-            receiveOnce: function (type, callback)
-            {
-                if (this.is_set(type))
-                    return;
-
-                this.receive(type, callback);
-            },
-            // Check if any callbacks assigned to a specified message type.
-            is_set: function (type)
-            {
-                return (undef != _callbacks[type]);
-            },
-            parent_url: function ()
-            {
-                return _parent_url;
-            },
-            parent_subdomain: function ()
-            {
-                return _parent_subdomain;
-            }
-        };
-    }();
-})(jQuery);
+!function(e,t){var s,n,o,i,r,a,c,p,u=this;u.FS=u.FS||{},u.FS.PostMessage=(n=new NoJQueryPostMessageMixin("postMessage","receiveMessage"),o={},i=decodeURIComponent(document.location.hash.replace(/^#/,"")),r=i.substring(0,i.indexOf("/","https://"===i.substring(0,8)?8:7)),a=""!==i,c=e(window),p=e("html"),{init:function(e,t){s=e,n.receiveMessage((function(e){var t=JSON.parse(e.data);if(o[t.type])for(var s=0;s<o[t.type].length;s++)o[t.type][s](t.data)}),s),FS.PostMessage.receiveOnce("forward",(function(e){window.location=e.url})),(t=t||[]).length>0&&c.on("scroll",(function(){for(var e=0;e<t.length;e++)FS.PostMessage.postScroll(t[e])}))},init_child:function(){this.init(r),e(window).bind("load",(function(){FS.PostMessage.postHeight(),FS.PostMessage.post("loaded")}))},hasParent:function(){return a},postHeight:function(t,s){t=t||0,s=s||"#wrap_section",this.post("height",{height:t+e(s).outerHeight(!0)})},postScroll:function(e){this.post("scroll",{top:c.scrollTop(),height:c.height()-parseFloat(p.css("paddingTop"))-parseFloat(p.css("marginTop"))},e)},post:function(e,t,s){console.debug("PostMessage.post",e),s?n.postMessage(JSON.stringify({type:e,data:t}),s.src,s.contentWindow):n.postMessage(JSON.stringify({type:e,data:t}),i,window.parent)},receive:function(e,s){console.debug("PostMessage.receive",e),t===o[e]&&(o[e]=[]),o[e].push(s)},receiveOnce:function(e,t){this.is_set(e)||this.receive(e,t)},is_set:function(e){return t!=o[e]},parent_url:function(){return i},parent_subdomain:function(){return r}})}(jQuery);
