+++ assets/scripts/index.php
@@ -1,0 +1,3 @@
+<?php
+	// Silence is golden.
+	// Hide file structure from users on unprotected servers.
