+++ assets/scss/admin/_switch.scss
@@ -1,3 +1,5 @@
+@use "sass:math";
+
 $switch-size: 24px;
 $switch-small-size: 0.75 * $switch-size;
 $switch-on-color: $wp-button-primary-background-color;
@@ -32,11 +34,10 @@ $switch-on-color: $wp-button-primary-background-color;
         top:           1px;
         width:         37px;
         height:        $switch-size + 1px;
-        border:        1px solid #ccc;
         border:        1px solid rgba(0, 0, 0, 0.3);
         border-radius: 4px;
-        background:    #fff;
-        @include linear-gradient(#ececec, #fff);
+        background-color:    #fff;
+        background-image: linear-gradient(to bottom, #ececec, #fff);
         box-shadow:    inset 0 1px 0 0 rgba(255, 255, 255, 0.5);
         z-index:       999;
         @include transition(0.40s cubic-bezier(0.54, 1.6, 0.5, 1));
@@ -54,7 +55,7 @@ $switch-on-color: $wp-button-primary-background-color;
     /* Round switch */
     &.fs-round
     {
-        top:     $switch-size / 3;
+        top:     math.div($switch-size, 3);
         padding: 4px ($switch-size + 1px);
         @include border-radius($switch-size);
