+++ package.json
@@ -9,18 +9,33 @@
   "main": "gulpfile.js",
   "scripts": {
     "translate": "gulp translate",
+    "build": "gulp build",
+    "dev": "gulp dev",
     "postinstall": "patch-package"
   },
   "repository": "Freemius/wordpress-sdk.git",
   "devDependencies": {
     "@transifex/api": "^7.0.1",
+    "autoprefixer": "^10.4.17",
+    "cssnano": "^6.0.3",
     "dotenv": "^16.4.1",
     "fancy-log": "^2.0.0",
     "gettext-parser": "^7.0.1",
     "gulp": "^4.0.2",
+    "gulp-concat": "^2.6.1",
+    "gulp-esbuild": "^0.12.0",
+    "gulp-livereload": "^4.0.2",
+    "gulp-plumber": "^1.2.1",
+    "gulp-postcss": "^9.1.0",
+    "gulp-sass": "^5.1.0",
     "gulp-sort": "^2.0.0",
+    "gulp-sourcemaps": "^3.0.0",
+    "gulp-terser": "^2.1.0",
     "gulp-wp-pot": "^2.5.0",
     "node-fetch": "^3.3.2",
-    "patch-package": "^8.0.0"
+    "node-sass-package-importer": "^5.3.3",
+    "patch-package": "^8.0.0",
+    "postcss": "^8.4.33",
+    "sass": "^1.70.0"
   }
 }
