+++ templates/powered-by.php
@@ -36,7 +36,7 @@
     if ( ! $fs->is_whitelabeled() && ! $fs->apply_filters( 'hide_freemius_powered_by', false ) ) {
         wp_enqueue_script( 'jquery' );
         wp_enqueue_script( 'json2' );
-        fs_enqueue_local_script( 'postmessage', 'nojquery.ba-postmessage.min.js' );
+        fs_enqueue_local_script( 'postmessage', 'nojquery.ba-postmessage.js' );
         fs_enqueue_local_script( 'fs-postmessage', 'postmessage.js' );
     ?>
 <div id="pframe"></div>
